refactor: remove code  duplication
