+++ apps/ledger-live-desktop/src/renderer/actions/UI.ts
@@ -1,7 +1,7 @@
 import { LiveAppManifest } from "@ledgerhq/live-common/platform/types";
 import { AppManifest } from "@ledgerhq/live-common/wallet-api/types";
 import { createAction } from "redux-actions";
-import { Data as CompleteExchangeData } from "~/renderer/modals/Platform/Exchange/CompleteExchange/Drawer";
+import { Data as CompleteExchangeData } from "~/renderer/modals/Platform/Exchange/CompleteExchange/Body";
 import { ExchangeType } from "@ledgerhq/live-common/wallet-api/react";
 
 export const openInformationCenter = createAction(
